requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
streamlit==1.35.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
groq==0.9.0
httpx==0.27.0
langchain==0.2.5
langchain-groq==0.1.6
langchain-community==0.2.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4